+++ src/main/kotlin/dev/jkcarino/revanced/patches/all/firebase/analytics/DeactivateFirebaseAnalyticsPatch.kt
@@ -1,38 +1,44 @@
-package dev.jkcarino.revanced.patches.all.webview
+package dev.jkcarino.revanced.patches.all.firebase.analytics
 
 import app.revanced.patcher.data.ResourceContext
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotation.Patch
-import dev.jkcarino.revanced.util.createElement
 import dev.jkcarino.revanced.util.asElementSequence
+import dev.jkcarino.revanced.util.createElement
 import dev.jkcarino.revanced.util.firstElementByTagName
 
 @Patch(
-    name = "Disable metrics collection in WebView",
-    description = "Disables the collection of diagnostic data or usage statistics that are uploaded to Google.",
+    name = "Deactivate Firebase Analytics",
+    description = "Deactivates Firebase Analytics and removes its associated broadcast receivers and services.",
+    dependencies = [
+        RemoveAdServicesPatch::class,
+        RemoveAdvertisingIdPatch::class,
+        RemoveAppMeasurementPatch::class,
+        RemoveGoogleAnalyticsPatch::class,
+    ],
     use = false
 )
 @Suppress("unused")
-object DisableWebViewMetricsPatch : ResourcePatch() {
+object DeactivateFirebaseAnalyticsPatch : ResourcePatch() {
     private const val META_DATA_TAG = "meta-data"
     private const val ATTRIBUTE_NAME = "android:name"
     private const val ATTRIBUTE_VALUE = "android:value"
-    private const val WEBVIEW_METRICS_OPT_OUT = "android.webkit.WebView.MetricsOptOut"
+    private const val FIREBASE_ANALYTICS_DEACTIVATED = "firebase_analytics_collection_deactivated"
 
     override fun execute(context: ResourceContext) {
         context.xmlEditor["AndroidManifest.xml"].use { editor ->
             val document = editor.file
             val application = document.firstElementByTagName("application")
 
-            val webViewMetricsOptOutMetaData = application.getElementsByTagName(META_DATA_TAG)
+            val firebaseAnalyticsMetaData = application.getElementsByTagName(META_DATA_TAG)
                 .asElementSequence()
-                .firstOrNull { it.getAttribute(ATTRIBUTE_NAME) == WEBVIEW_METRICS_OPT_OUT }
+                .firstOrNull { it.getAttribute(ATTRIBUTE_NAME) == FIREBASE_ANALYTICS_DEACTIVATED }
                 ?.setAttribute(ATTRIBUTE_VALUE, "true")
 
-            if (webViewMetricsOptOutMetaData == null) {
+            if (firebaseAnalyticsMetaData == null) {
                 application.appendChild(
                     document.createElement(META_DATA_TAG) {
-                        setAttribute(ATTRIBUTE_NAME, WEBVIEW_METRICS_OPT_OUT)
+                        setAttribute(ATTRIBUTE_NAME, FIREBASE_ANALYTICS_DEACTIVATED)
                         setAttribute(ATTRIBUTE_VALUE, "true")
                     }
                 )