+++ src/main/kotlin/dev/jkcarino/revanced/patches/all/firebase/performance/DeactivateFirebasePerfPatch.kt
@@ -1,38 +1,38 @@
-package dev.jkcarino.revanced.patches.all.webview
+package dev.jkcarino.revanced.patches.all.firebase.performance
 
 import app.revanced.patcher.data.ResourceContext
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotation.Patch
-import dev.jkcarino.revanced.util.createElement
 import dev.jkcarino.revanced.util.asElementSequence
+import dev.jkcarino.revanced.util.createElement
 import dev.jkcarino.revanced.util.firstElementByTagName
 
 @Patch(
-    name = "Disable metrics collection in WebView",
-    description = "Disables the collection of diagnostic data or usage statistics that are uploaded to Google.",
+    name = "Deactivate Firebase Performance Monitoring",
+    description = "Deactivates the collection of performance data on app start up time, network requests, and other related metrics.",
     use = false
 )
 @Suppress("unused")
-object DisableWebViewMetricsPatch : ResourcePatch() {
+object DeactivateFirebasePerfPatch : ResourcePatch() {
     private const val META_DATA_TAG = "meta-data"
     private const val ATTRIBUTE_NAME = "android:name"
     private const val ATTRIBUTE_VALUE = "android:value"
-    private const val WEBVIEW_METRICS_OPT_OUT = "android.webkit.WebView.MetricsOptOut"
+    private const val FIREBASE_PERF_DEACTIVATED = "firebase_performance_collection_deactivated"
 
     override fun execute(context: ResourceContext) {
         context.xmlEditor["AndroidManifest.xml"].use { editor ->
             val document = editor.file
             val application = document.firstElementByTagName("application")
 
-            val webViewMetricsOptOutMetaData = application.getElementsByTagName(META_DATA_TAG)
+            val firebasePerfMetaData = application.getElementsByTagName(META_DATA_TAG)
                 .asElementSequence()
-                .firstOrNull { it.getAttribute(ATTRIBUTE_NAME) == WEBVIEW_METRICS_OPT_OUT }
+                .firstOrNull { it.getAttribute(ATTRIBUTE_NAME) == FIREBASE_PERF_DEACTIVATED }
                 ?.setAttribute(ATTRIBUTE_VALUE, "true")
 
-            if (webViewMetricsOptOutMetaData == null) {
+            if (firebasePerfMetaData == null) {
                 application.appendChild(
                     document.createElement(META_DATA_TAG) {
-                        setAttribute(ATTRIBUTE_NAME, WEBVIEW_METRICS_OPT_OUT)
+                        setAttribute(ATTRIBUTE_NAME, FIREBASE_PERF_DEACTIVATED)
                         setAttribute(ATTRIBUTE_VALUE, "true")
                     }
                 )