+++ src/main/kotlin/dev/jkcarino/revanced/patches/all/webview/DisableWebViewSafeBrowsingPatch.kt
@@ -3,7 +3,9 @@ package dev.jkcarino.revanced.patches.all.webview
 import app.revanced.patcher.data.ResourceContext
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotation.Patch
-import org.w3c.dom.Element
+import dev.jkcarino.revanced.util.createElement
+import dev.jkcarino.revanced.util.asElementSequence
+import dev.jkcarino.revanced.util.firstElementByTagName
 
 @Patch(
     name = "Disable Google Safe Browsing in WebView",
@@ -20,23 +22,21 @@ object DisableWebViewSafeBrowsingPatch : ResourcePatch() {
     override fun execute(context: ResourceContext) {
         context.xmlEditor["AndroidManifest.xml"].use { editor ->
             val document = editor.file
-            val applicationTag = document
-                .getElementsByTagName("application")
-                .item(0) as Element
-            val metaData = applicationTag.getElementsByTagName(META_DATA_TAG)
+            val application = document.firstElementByTagName("application")
 
-            for (i in 0 until metaData.length) {
-                val element = metaData.item(i) as? Element ?: continue
-                if (element.getAttribute(ATTRIBUTE_NAME) == WEBVIEW_SAFE_BROWSING) {
-                    element.setAttribute(ATTRIBUTE_VALUE, "false")
-                    return
-                }
+            val webViewSafeBrowsingMetaData = application.getElementsByTagName(META_DATA_TAG)
+                .asElementSequence()
+                .firstOrNull { it.getAttribute(ATTRIBUTE_NAME) == WEBVIEW_SAFE_BROWSING }
+                ?.setAttribute(ATTRIBUTE_VALUE, "false")
+
+            if (webViewSafeBrowsingMetaData == null) {
+                application.appendChild(
+                    document.createElement(META_DATA_TAG) {
+                        setAttribute(ATTRIBUTE_NAME, WEBVIEW_SAFE_BROWSING)
+                        setAttribute(ATTRIBUTE_VALUE, "false")
+                    }
+                )
             }
-
-            document.createElement(META_DATA_TAG).apply {
-                setAttribute(ATTRIBUTE_NAME, WEBVIEW_SAFE_BROWSING)
-                setAttribute(ATTRIBUTE_VALUE, "false")
-            }.also(applicationTag::appendChild)
         }
     }
 }